[#228] use multi-release JARs
Really fixes #228.
This will also show the NPE everyone is talking about in #91.
